requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
cryptography==41.0.7
gunicorn>=21,<24
psycopg[binary]>=3.1,<4
